fix : 나무 상세 페이지 로직 수정
